Move ReservedListDao related classes to schema package
