+++ core/src/test/java/google/registry/tools/CreateOrUpdateReservedListCommandTestCase.java
@@ -25,9 +25,9 @@ import com.beust.jcommander.ParameterException;
 import com.google.common.io.Files;
 import com.google.common.truth.Truth8;
 import google.registry.model.registry.label.ReservedList;
-import google.registry.model.registry.label.ReservedListDao;
 import google.registry.model.transaction.JpaTransactionManagerRule;
 import google.registry.schema.tld.ReservedList.ReservedEntry;
+import google.registry.schema.tld.ReservedListDao;
 import java.io.File;
 import java.io.IOException;
 import java.util.Map;
