+++ core/src/test/java/google/registry/tools/CreateReservedListCommandTest.java
@@ -27,8 +27,8 @@ import static org.joda.time.DateTimeZone.UTC;
 import com.google.common.collect.ImmutableMap;
 import google.registry.model.registry.Registry;
 import google.registry.model.registry.label.ReservedList;
-import google.registry.model.registry.label.ReservedListDao;
 import google.registry.schema.tld.ReservedList.ReservedEntry;
+import google.registry.schema.tld.ReservedListDao;
 import org.joda.time.DateTime;
 import org.junit.Before;
 import org.junit.Test;
